feat(data): now saves a proper nested json using path_provider instead of shared_preferences

+++ lib/screens/resonator_detail_screen.dart
@@ -21,7 +21,7 @@ class ResonatorDetailScreen extends StatefulWidget {
 
 class _ResonatorDetailScreenState extends State<ResonatorDetailScreen> {
   String energyBuff = 'None'; // None, Yangyang, Zhezhi
-  int totalER = 100;
+  double totalER = 100.0;
   late List<Map<String, double>> echoStats;
   bool loading = false;
   EchoSet? lastResult;
